+++ src/Index.ts
@@ -1,2 +1,3 @@
 export * from "./ChangeChecker";
 export * from "./DiffTypes";
+export * from "./Errors";